test: move fixtures to test files where they're used
