+++ tests/utils/test_copy_folder_to_container.py
@@ -1,7 +1,24 @@
 from pathlib import Path
+from typing import Generator, cast
 
-from tests.conftest import AlpineContainer
+import pytest
+
+from tomodachi_testcontainers import DockerContainer
 from tomodachi_testcontainers.utils import copy_folder_to_container
+
+
+class AlpineContainer(DockerContainer):
+    def __init__(self) -> None:
+        super().__init__(image="alpine:latest")
+
+    def log_message_on_container_start(self) -> str:
+        return "Alpine container started"
+
+
+@pytest.fixture(scope="module")
+def alpine_container() -> Generator[AlpineContainer, None, None]:
+    with AlpineContainer().with_command("sleep infinity") as container:
+        yield cast(AlpineContainer, container)
 
 
 def test_copy_folder_to_container(alpine_container: AlpineContainer) -> None:
